refactor to be seperate component, and fix the state.
